requirements.txt
langchain
langchain-openai
mysql-connector-python
